refactor(shared): break env facade import cycle
_env_paths.py and _env_db.py imported _sanitized_env_path back from
_env.py, which imports them at its own bottom -- a real two-way reference
that only worked because _env.py happened to initiate the cycle first
every time. A direct `import quodeq.shared._env_paths` in isolation would
have hit a partially-initialized _env module.

Moved _sanitized_env_path into a new leaf module, _env_sanitize.py, with
no dependents of its own. _env_paths.py and _env_db.py now import it from
there instead of from _env.py; _env.py imports it from there too and
re-exports it, so the module's public surface is unchanged. Chose the
leaf-module extraction over deferred per-function imports: 6 call sites
across two files would each need a local import, versus one leaf module
both sides import cleanly, and it matches the existing shared/ pattern of
small single-purpose leaf modules (_config.py, _repo.py, _diff.py, etc.).

No behavior change; no test edits.
