+++ src/quodeq/shared/_env_db.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 import os
 from pathlib import Path
 
-from quodeq.shared._env import _sanitized_env_path
+from quodeq.shared._env_sanitize import _sanitized_env_path
 
 _SQLITE_DISABLE_TRUTHY = {"1", "true", "yes", "on"}
